Add delete api

+++ pkg/db/ent/recommend.go
@@ -25,11 +25,11 @@ type Recommend struct {
 	// Message holds the value of the "message" field.
 	Message string `json:"message,omitempty"`
 	// CreateAt holds the value of the "create_at" field.
-	CreateAt int64 `json:"create_at,omitempty"`
+	CreateAt uint32 `json:"create_at,omitempty"`
 	// UpdateAt holds the value of the "update_at" field.
-	UpdateAt int64 `json:"update_at,omitempty"`
+	UpdateAt uint32 `json:"update_at,omitempty"`
 	// DeleteAt holds the value of the "delete_at" field.
-	DeleteAt int64 `json:"delete_at,omitempty"`
+	DeleteAt uint32 `json:"delete_at,omitempty"`
 }
 
 // scanValues returns the types for scanning values from sql.Rows.
@@ -92,19 +92,19 @@ func (r *Recommend) assignValues(columns []string, values []interface{}) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field create_at", values[i])
 			} else if value.Valid {
-				r.CreateAt = value.Int64
+				r.CreateAt = uint32(value.Int64)
 			}
 		case recommend.FieldUpdateAt:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field update_at", values[i])
 			} else if value.Valid {
-				r.UpdateAt = value.Int64
+				r.UpdateAt = uint32(value.Int64)
 			}
 		case recommend.FieldDeleteAt:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field delete_at", values[i])
 			} else if value.Valid {
-				r.DeleteAt = value.Int64
+				r.DeleteAt = uint32(value.Int64)
 			}
 		}
 	}
