+++ pkg/db/ent/recommend/where.go
@@ -120,21 +120,21 @@ func Message(v string) predicate.Recommend {
 }
 
 // CreateAt applies equality check predicate on the "create_at" field. It's identical to CreateAtEQ.
-func CreateAt(v int64) predicate.Recommend {
+func CreateAt(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldCreateAt), v))
 	})
 }
 
 // UpdateAt applies equality check predicate on the "update_at" field. It's identical to UpdateAtEQ.
-func UpdateAt(v int64) predicate.Recommend {
+func UpdateAt(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldUpdateAt), v))
 	})
 }
 
 // DeleteAt applies equality check predicate on the "delete_at" field. It's identical to DeleteAtEQ.
-func DeleteAt(v int64) predicate.Recommend {
+func DeleteAt(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldDeleteAt), v))
 	})
@@ -480,21 +480,21 @@ func MessageContainsFold(v string) predicate.Recommend {
 }
 
 // CreateAtEQ applies the EQ predicate on the "create_at" field.
-func CreateAtEQ(v int64) predicate.Recommend {
+func CreateAtEQ(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtNEQ applies the NEQ predicate on the "create_at" field.
-func CreateAtNEQ(v int64) predicate.Recommend {
+func CreateAtNEQ(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtIn applies the In predicate on the "create_at" field.
-func CreateAtIn(vs ...int64) predicate.Recommend {
+func CreateAtIn(vs ...uint32) predicate.Recommend {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -511,7 +511,7 @@ func CreateAtIn(vs ...int64) predicate.Recommend {
 }
 
 // CreateAtNotIn applies the NotIn predicate on the "create_at" field.
-func CreateAtNotIn(vs ...int64) predicate.Recommend {
+func CreateAtNotIn(vs ...uint32) predicate.Recommend {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -528,49 +528,49 @@ func CreateAtNotIn(vs ...int64) predicate.Recommend {
 }
 
 // CreateAtGT applies the GT predicate on the "create_at" field.
-func CreateAtGT(v int64) predicate.Recommend {
+func CreateAtGT(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtGTE applies the GTE predicate on the "create_at" field.
-func CreateAtGTE(v int64) predicate.Recommend {
+func CreateAtGTE(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtLT applies the LT predicate on the "create_at" field.
-func CreateAtLT(v int64) predicate.Recommend {
+func CreateAtLT(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtLTE applies the LTE predicate on the "create_at" field.
-func CreateAtLTE(v int64) predicate.Recommend {
+func CreateAtLTE(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldCreateAt), v))
 	})
 }
 
 // UpdateAtEQ applies the EQ predicate on the "update_at" field.
-func UpdateAtEQ(v int64) predicate.Recommend {
+func UpdateAtEQ(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldUpdateAt), v))
 	})
 }
 
 // UpdateAtNEQ applies the NEQ predicate on the "update_at" field.
-func UpdateAtNEQ(v int64) predicate.Recommend {
+func UpdateAtNEQ(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldUpdateAt), v))
 	})
 }
 
 // UpdateAtIn applies the In predicate on the "update_at" field.
-func UpdateAtIn(vs ...int64) predicate.Recommend {
+func UpdateAtIn(vs ...uint32) predicate.Recommend {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -587,7 +587,7 @@ func UpdateAtIn(vs ...int64) predicate.Recommend {
 }
 
 // UpdateAtNotIn applies the NotIn predicate on the "update_at" field.
-func UpdateAtNotIn(vs ...int64) predicate.Recommend {
+func UpdateAtNotIn(vs ...uint32) predicate.Recommend {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -604,49 +604,49 @@ func UpdateAtNotIn(vs ...int64) predicate.Recommend {
 }
 
 // UpdateAtGT applies the GT predicate on the "update_at" field.
-func UpdateAtGT(v int64) predicate.Recommend {
+func UpdateAtGT(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldUpdateAt), v))
 	})
 }
 
 // UpdateAtGTE applies the GTE predicate on the "update_at" field.
-func UpdateAtGTE(v int64) predicate.Recommend {
+func UpdateAtGTE(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldUpdateAt), v))
 	})
 }
 
 // UpdateAtLT applies the LT predicate on the "update_at" field.
-func UpdateAtLT(v int64) predicate.Recommend {
+func UpdateAtLT(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldUpdateAt), v))
 	})
 }
 
 // UpdateAtLTE applies the LTE predicate on the "update_at" field.
-func UpdateAtLTE(v int64) predicate.Recommend {
+func UpdateAtLTE(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldUpdateAt), v))
 	})
 }
 
 // DeleteAtEQ applies the EQ predicate on the "delete_at" field.
-func DeleteAtEQ(v int64) predicate.Recommend {
+func DeleteAtEQ(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldDeleteAt), v))
 	})
 }
 
 // DeleteAtNEQ applies the NEQ predicate on the "delete_at" field.
-func DeleteAtNEQ(v int64) predicate.Recommend {
+func DeleteAtNEQ(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldDeleteAt), v))
 	})
 }
 
 // DeleteAtIn applies the In predicate on the "delete_at" field.
-func DeleteAtIn(vs ...int64) predicate.Recommend {
+func DeleteAtIn(vs ...uint32) predicate.Recommend {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -663,7 +663,7 @@ func DeleteAtIn(vs ...int64) predicate.Recommend {
 }
 
 // DeleteAtNotIn applies the NotIn predicate on the "delete_at" field.
-func DeleteAtNotIn(vs ...int64) predicate.Recommend {
+func DeleteAtNotIn(vs ...uint32) predicate.Recommend {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -680,28 +680,28 @@ func DeleteAtNotIn(vs ...int64) predicate.Recommend {
 }
 
 // DeleteAtGT applies the GT predicate on the "delete_at" field.
-func DeleteAtGT(v int64) predicate.Recommend {
+func DeleteAtGT(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldDeleteAt), v))
 	})
 }
 
 // DeleteAtGTE applies the GTE predicate on the "delete_at" field.
-func DeleteAtGTE(v int64) predicate.Recommend {
+func DeleteAtGTE(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldDeleteAt), v))
 	})
 }
 
 // DeleteAtLT applies the LT predicate on the "delete_at" field.
-func DeleteAtLT(v int64) predicate.Recommend {
+func DeleteAtLT(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldDeleteAt), v))
 	})
 }
 
 // DeleteAtLTE applies the LTE predicate on the "delete_at" field.
-func DeleteAtLTE(v int64) predicate.Recommend {
+func DeleteAtLTE(v uint32) predicate.Recommend {
 	return predicate.Recommend(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldDeleteAt), v))
 	})
